requirements.txt
openai
ollama
icrawler
python-pptx
python-dotenv
argparse